+++ src/components/Card/CHANGELOG.md
@@ -4,3 +4,5 @@
 - 2025-11-25: Add the sub-component `Card.ContentInScrollArea`, and support the
   `withBorder` prop for `Card.Header` and `Card.Footer`. Some styles are updated
   to make it looks better.
+- 2025-12-07: Add the `withTypography` prop to `Card.Content` and
+  `Card.ContentInScrollArea`.